fix(test): Avoid polluting common namespace
Since RSpec uses just a block to define test cases,
constants defined in the block are accessible from all other files.
So we need to avoid defining constants in test files.

Also, a sample class in integration tests must be defined inside of
`it` blocks because if it uses `injectable` method, it changes
Dio's global state outside of test cases. This can't be reset by
`before(:each)` hook.